\name{getDf}
\docType{methods}
\alias{getDf-methods}
\alias{getDf,WeatherStation-method}
\title{ Get dataframe frome a S4 object \code{getDf}  }
\description{
 Get data frame from S4 object \code{getDf}  
}
\section{Methods}{
\describe{

\item{\code{signature(x = "WeatherStation")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}

